Split organization filters to separate file
